GO terms in diamond plot

Visualize GO enrichment result in a diamond plot

GO_diamondplot(data, palette = TRUE, output_name)

Arguments

data
A tibble, with following columns

annotated: numeric, number of genes enriched in the GO-term

total: numeric, total number of genes assigned to the GO-term

pvalue: numeric, pvalue of enrichment

category: character, a group-name to annotate the GO and color the diamonds

class: character, optional if there are two categories, second category should be under column class.

palette
Character, to choose from available palette or define a palette.default:TRUE

two_color: a palette of pink and green

three_color: a palette of orange, pink and green

NULL: user input of character vector of colors

TRUE: a palette of colors in rainbow

output_name
A character vector containing name of the output plot.default:Sample

Value

a diamond GO plot and image file of same.

Examples

if (FALSE) {

data <- readr::read_delim(system.file("extdata/genesets/an_af_pm_spores_GO.txt" , package = "FungalSporeAnalysis"), col_names = TRUE, delim="\t")
GO_diamondplot(data, output_name = "An_Af_Pm_GO", palette="three_color")

}