plot correlation scatterplot for bw files

correlation between two replicates of samples with correlation value.

bw_corr(dat, multiBigwigSummary_output = TRUE, pattern = NULL)

Arguments

dat
Absolute path of count matrix. The input file can be generated by deeptools multiBigwigSummary bins command. The output should be a counts matrix --outRawCounts with first three columns:chr start end and later sample wise counts. Alternatively, a matrix of expression values with first column gene-name and later expression values.

multiBigwigSummary_output
Logical, whether the input matrix is multiBigwigSummary bins out put or not. default:TRUE

pattern
A character vector, defining pattern to split the columns in replicates. default: E.g. sampleA_set1, sampleA_set2, sampleB_set1, sampleB_set2. pattern="(.*)_(.*)". will have groups of sample: sampleA, sampleB and replicate: set1, set2

Value

a multi-sample scatterplot of log2 values with correlation value

Examples

if (FALSE) {

dat_path <- system.file("extdata//correlation_data/an_histone_spore_count_mat.tab" , package = "FungalSporeAnalysis")
dat <- read.delim(dat_path,sep="\t", header = TRUE)
gp <- bw_corr(dat, pattern = ".*_spore_(.*)_(.*)")
ggsave(gp, filename = "An_spore_histone_bwscatterplot.png", height=5, width=5, device = "png", units = "in")

}