profile plot specific to a genelist

Profile plot for list of genes to be subsetted from genome feature file.

genelist_specific_profileplot(feature_txDb, bw_files, genelist = NULL,
  tss = TRUE, palette = "cream_brown", max_key = 6, min_key = 0,
  top_line = TRUE, ymax = 6, ymin = 1, log2 = TRUE,
  rename = FALSE, output = TRUE, output_name = "Sample")

Arguments

feature_txDb
A TxDb object of gene feature file, generated by GenomicFeatures::makeTxDbFromGFF()

bw_files
A character vector containing either name or names of GRanges object of bw or bdg files. GRanges object for bw_files can be generated by: rtracklayer::import()

genelist
A tibble with two columns 1. gene_id and 2.Expression value without header, to subset genes from the feature file. default:NULL

tss
Logical, if TRUE plot 1kb up/downstream of start co-ordinate; if FALSE plot 1kb up/downstream of gene-body default: TRUE

palette
A character vector to choose sequential palette;

white_green

white_red

white_blue

cream_pink

cream_brown

cream_green

default: cream_brown

max_key
Numeric, maximum limit of the color key.default:6

min_key
Numeric, mimimum limit of the color key.default:0

top_line
Logical, whether to plot profile with average signal as line plot at the top of plot. default: TRUE

ymax
Numeric, y-axis maximum limit for lineplot default:6

ymin
Numeric, y-axis minimum limit for lineplot default:1

log2
Logical, whether to plot profile as log-transformed or not.

rename
Logical, whether to remove characters after first - OR _, suitable for long names of bw objects.default: FALSE

output
Logical, to return a output of profile plot or not. default: TRUE, with output_name,if FALSE; plot will be returned ** Highly recommended to plot in outfile when .bw files are more than 2.

output_name
A character vector containing name of the output plot.default:Sample

Value

Parallel plots of profiles(log2) for given gene lists, ordered by given expression value (high to low).

Examples

if (FALSE) {

 # Load feature file from system data, to run example data
 feature_txDb <- AnnotationDbi::loadDb(system.file("extdata/sqllite/an_feature_file_s10_m04_r07.sqlite" , package = "FungalSporeAnalysis"))

 # Or load feature file as TxDb
 feature_txDb <- GenomicFeatures::makeTxDbFromGFF("an_feature_file_s10_m04_r07.gff")

 # Load bw files as GRanges object
 pol2_veA_wt_spore <- rtracklayer::import.bw("pol2_veA_wt_spore_mix22_CACAGTTGGT_normalized_repeat.bw")
 TBP_veA_wt_spore <- rtracklayer::import.bw("TBP_veA_wt_spore_mix22_CAGTTGGT_normalized_repeat.bw")

 # Load GRanges object from system data, to run example data
 data_files <- system.file("extdata/sysdata.rda" , package = "FungalSporeAnalysis")
 load(data_files)

 # Load example gene_list file
 gene_list <- readr::read_delim(system.file("extdata/genesets/an_spore_pol2.txt" , package = "FungalSporeAnalysis"), delim="\t", col_names = FALSE)

 #################

 bw_files <- c("pol2_veA_wt_spore","TBP_veA_wt_spore")
 genelist_specific_profileplot(feature_txDb=feature_txDb,bw_files = bw_files, genelist=gene_list, output_name="An_Pol2_TBP_TFIIB", ymin=3)

}