heatmap of gene expressions by class

Heatmap of genes expression value categorized by classes

ggplot_heatmap(input_data, threshold = FALSE, output_name = "Sample")

Arguments

input_data
A tibble, with four columns. First column name should be genes and second column name should be class. Other columns can have unique names to plot.

threshold
Logical, to set maximum limit of the color key.default:FALSE. If TRUE on-run threshold value should be provided.

output_name
A character vector, containing name of the output plot.default:Sample

Value

heatmap of expression of given genes categorized by class and output image file of same.

Examples

if (FALSE) {

dat <- readr::read_delim(system.file("extdata/genesets/an_spore_hypha_specificgenes.txt" , package = "FungalSporeAnalysis"), delim="\t", col_names = TRUE)
input_data <- dat %>% dplyr::filter(class=="spore_maturation")
ggplot_heatmap(input_data ,threshold = TRUE, output_name = "plots/An_spores_maturation_genes_exprsn")

}