profile plots of signal normalized to control

Control normalized signal are plotted for all genes or specific list of genes.

profiles_normalized_by_control(feature_txDb, bw_test, bw_control,
  genelist_1 = NULL, genelist_2 = NULL, ymax = 6, ymin = 1,
  output_name = "Sample", top_line = TRUE)

Arguments

feature_txDb
A TxDb object of gene feature file, generated by GenomicFeatures::makeTxDbFromGFF()

bw_test
A character vector containing name of the GRanges object of bw file.GRanges object for bw_files can be generated by: rtracklayer::import()

bw_control
A character vector containing name of the GRanges object of bw file to normalize the bw_test. GRanges object for bw_files can be generated by: rtracklayer::import()

genelist_1
A tibble with two columns 1. gene_id and 2.Expression value without header, to subset genes from the feature file. default:NULL

genelist_2
A tibble with either one or two columns to be considered as control genes. 1. gene_id and/or 2.Expression value without header, to subset genes from the feature file. default:NULL

ymax
Numeric, y-axis max limit for lineplot default:6

ymin
Numeric, y-axis minimum limit for lineplot default:1

output_name
A character vector containing name of the output plot.default:Sample

top_line
Logical, whether to plot profile with average signal as line plot at the top of plot. default: TRUE

Value

An image of two heatmaps normalized to control bw.

Heatmap_list1: profile of genelist_1 ordered by expression values (high to low).

Heatmap_list1: profile of genelist_2.

Examples

if (FALSE) {

 # Load feature file from system data, to run example data
 feature_txDb <- AnnotationDbi::loadDb(system.file("extdata/sqllite/an_feature_file_s10_m04_r07.sqlite" , package = "FungalSporeAnalysis"))

 # Or load feature file as TxDb
 feature_txDb <- GenomicFeatures::makeTxDbFromGFF("an_feature_file_s10_m04_r07.gff")

 # Load bw files as GRanges object
 H3AC_veA_wt_spore <- rtracklayer::import.bw("H3AC_veA_wt_spore_mix22_CACAGTTGGT_normalized_repeat.bw")
 H3_an_spore <- rtracklayer::import.bw("H3_an_spore_spore_mix22_CACAGTTGGT_normalized_repeat.bw")

 # Load GRanges object from system data, to run example data
 data_files <- system.file("extdata/sysdata.rda" , package = "FungalSporeAnalysis")
 load(data_files)

 # Load gene_lists for specific genes
 genelist_1 <- readr::read_delim(system.file("extdata/genesets/an_spore_pol2.txt" , package = "FungalSporeAnalysis"), delim="\t", col_names = FALSE)
 genelist_2 <- readr::read_delim(system.file("extdata/genesets/an_spore_pol2_controlgenes.txt", package = "FungalSporeAnalysis"), delim="\t", col_names = FALSE)

 #################################
 profiles_normalized_by_control(feature_txDb = feature_txDb,bw_test = "H3AC_veA_wt_spore", bw_control = "H3_an_spore",genelist_1 = genelist_1,genelist_2 = genelist_2, ymax = 3.8,ymin=0.5, output_name = "H3Ac_veA_wt_spore")

}